# docs/swagger-inline.yaml
openapi: 3.0.0
paths:
...
  /users:
    post:
      operationId: UsersController_create
      summary: 创建用户
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                username:
                  type: string
                  description: 用户名
                  example: zhangsan
                password:
                  type: string
                  description: 密码
                  example: '123456'
                name:
                  type: string
                  description: 真实姓名
                  example: 张三
                email:
                  type: string
                  description: 邮箱
                  example: zhangsan@example.com
                phone:
                  type: string
                  description: 电话
                  example: '13800138000'
                position:
                  type: string
                  description: 职位
                  example: 前端工程师
                department_id:
                  type: number
                  description: 部门ID
                  example: 1
                leader_id:
                  type: number
                  description: 直属领导ID
                  example: 3
                role_ids:
                  description: 角色ID列表
                  example:
                    - 4
                  type: array
                  items:
                    type: string
                join_date:
                  type: string
                  description: 入职日期
                  example: '2023-03-15'
              required:
                - username
                - password
                - name
                - role_ids
      responses:
        '201':
          description: 创建成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: number
                    description: 用户ID
                    example: 1
                  username:
                    type: string
                    description: 用户名
                    example: admin
                  name:
                    type: string
                    description: 姓名
                    example: 系统管理员
                  email:
                    type: string
                    description: 邮箱
                    example: admin@example.com
                  phone:
                    type: string
                    description: 电话
                    example: '13800138000'
                  avatar:
                    type: string
                    description: 头像
                    example: avatar.jpg
                  status:
                    type: number
                    description: 状态
                    example: 1
                  join_date:
                    format: date-time
                    type: string
                    description: 入职日期
                    example: '2023-01-01'
                  position:
                    type: string
                    description: 职位
                    example: 系统管理员
                  department:
                    description: 部门信息
                    allOf:
                      - type: object
                        properties:
                          id:
                            type: number
                            description: 部门ID
                            example: 1
                          name:
                            type: string
                            description: 部门名称
                            example: 技术部
                          description:
                            type: string
                            description: 部门描述
                            example: 负责技术研发工作
                          parent_id:
                            type: number
                            description: 父部门ID
                            example: 1
                          sort_order:
                            type: number
                            description: 排序
                            example: 0
                          status:
                            type: number
                            description: 状态
                            example: 1
                          employeeCount:
                            type: number
                            description: 员工数量
                            example: 10
                          created_at:
                            format: date-time
                            type: string
                            description: 创建时间
                            example: '2023-01-01T00:00:00Z'
                          updated_at:
                            format: date-time
                            type: string
                            description: 更新时间
                            example: '2023-01-01T00:00:00Z'
                        required:
                          - id
                          - name
                          - sort_order
                          - status
                          - created_at
                          - updated_at
                  leader:
                    description: 直属领导
                    allOf:
                      - type: object
                        properties:
                          id:
                            type: number
                            description: 用户ID
                            example: 1
                          username:
                            type: string
                            description: 用户名
                            example: admin
                          name:
                            type: string
                            description: 姓名
                            example: 系统管理员
                          position:
                            type: string
                            description: 职位
                            example: 系统管理员
                        required:
                          - id
                          - username
                          - name
                  roles:
                    description: 角色列表
                    type: array
                    items:
                      type: object
                      properties:
                        id:
                          type: number
                          description: 角色ID
                          example: 1
                        code:
                          type: string
                          description: 角色代码
                          example: admin
                        name:
                          type: string
                          description: 角色名称
                          example: 系统管理员
                        description:
                          type: string
                          description: 角色描述
                          example: 系统管理员角色
                        permissions:
                          description: 权限列表
                          example:
                            - user:read
                            - user:write
                          type: array
                          items:
                            type: string
                        status:
                          type: number
                          description: 状态
                          example: 1
                        created_at:
                          format: date-time
                          type: string
                          description: 创建时间
                          example: '2023-01-01T00:00:00Z'
                        updated_at:
                          format: date-time
                          type: string
                          description: 更新时间
                          example: '2023-01-01T00:00:00Z'
                      required:
                        - id
                        - code
                        - name
                        - status
                        - created_at
                        - updated_at
                  created_at:
                    format: date-time
                    type: string
                    description: 创建时间
                    example: '2023-01-01T00:00:00Z'
                  updated_at:
                    format: date-time
                    type: string
                    description: 更新时间
                    example: '2023-01-01T00:00:00Z'
                required:
                  - id
                  - username
                  - name
                  - status
                  - roles
                  - created_at
                  - updated_at
        '400':
          description: 参数错误
        '401':
          description: 未授权
        '403':
          description: 无权限
      tags:
        - 用户管理
      security:
        - bearer: []
    get:
      operationId: UsersController_findAll
      summary: 获取用户列表
      parameters:
        - name: page
          required: false
          in: query
          description: 页码
          schema:
            example: 1
            type: number
        - name: limit
          required: false
          in: query
          description: 每页数量
          schema:
            example: 10
            type: number
        - name: department_id
          required: false
          in: query
          description: 部门ID
          schema:
            example: 1
            type: number
        - name: role
          required: false
          in: query
          description: 角色
          schema:
            example: employee
            type: string
        - name: search
          required: false
          in: query
          description: 搜索关键词
          schema:
            example: 张三
            type: string
      responses:
        '200':
          description: 获取成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    description: 用户列表
                    type: array
                    items:
                      type: object
                      properties:
                        id:
                          type: number
                          description: 用户ID
                          example: 1
                        username:
                          type: string
                          description: 用户名
                          example: admin
                        name:
                          type: string
                          description: 姓名
                          example: 系统管理员
                        email:
                          type: string
                          description: 邮箱
                          example: admin@example.com
                        phone:
                          type: string
                          description: 电话
                          example: '13800138000'
                        avatar:
                          type: string
                          description: 头像
                          example: avatar.jpg
                        status:
                          type: number
                          description: 状态
                          example: 1
                        join_date:
                          format: date-time
                          type: string
                          description: 入职日期
                          example: '2023-01-01'
                        position:
                          type: string
                          description: 职位
                          example: 系统管理员
                        department:
                          description: 部门信息
                          allOf:
                            - type: object
                              properties:
                                id:
                                  type: number
                                  description: 部门ID
                                  example: 1
                                name:
                                  type: string
                                  description: 部门名称
                                  example: 技术部
                                description:
                                  type: string
                                  description: 部门描述
                                  example: 负责技术研发工作
                                parent_id:
                                  type: number
                                  description: 父部门ID
                                  example: 1
                                sort_order:
                                  type: number
                                  description: 排序
                                  example: 0
                                status:
                                  type: number
                                  description: 状态
                                  example: 1
                                employeeCount:
                                  type: number
                                  description: 员工数量
                                  example: 10
                                created_at:
                                  format: date-time
                                  type: string
                                  description: 创建时间
                                  example: '2023-01-01T00:00:00Z'
                                updated_at:
                                  format: date-time
                                  type: string
                                  description: 更新时间
                                  example: '2023-01-01T00:00:00Z'
                              required:
                                - id
                                - name
                                - sort_order
                                - status
                                - created_at
                                - updated_at
                        leader:
                          description: 直属领导
                          allOf:
                            - type: object
                              properties:
                                id:
                                  type: number
                                  description: 用户ID
                                  example: 1
                                username:
                                  type: string
                                  description: 用户名
                                  example: admin
                                name:
                                  type: string
                                  description: 姓名
                                  example: 系统管理员
                                position:
                                  type: string
                                  description: 职位
                                  example: 系统管理员
                              required:
                                - id
                                - username
                                - name
                        roles:
                          description: 角色列表
                          type: array
                          items:
                            type: object
                            properties:
                              id:
                                type: number
                                description: 角色ID
                                example: 1
                              code:
                                type: string
                                description: 角色代码
                                example: admin
                              name:
                                type: string
                                description: 角色名称
                                example: 系统管理员
                              description:
                                type: string
                                description: 角色描述
                                example: 系统管理员角色
                              permissions:
                                description: 权限列表
                                example:
                                  - user:read
                                  - user:write
                                type: array
                                items:
                                  type: string
                              status:
                                type: number
                                description: 状态
                                example: 1
                              created_at:
                                format: date-time
                                type: string
                                description: 创建时间
                                example: '2023-01-01T00:00:00Z'
                              updated_at:
                                format: date-time
                                type: string
                                description: 更新时间
                                example: '2023-01-01T00:00:00Z'
                            required:
                              - id
                              - code
                              - name
                              - status
                              - created_at
                              - updated_at
                        created_at:
                          format: date-time
                          type: string
                          description: 创建时间
                          example: '2023-01-01T00:00:00Z'
                        updated_at:
                          format: date-time
                          type: string
                          description: 更新时间
                          example: '2023-01-01T00:00:00Z'
                      required:
                        - id
                        - username
                        - name
                        - status
                        - roles
                        - created_at
                        - updated_at
                  total:
                    type: number
                    description: 总记录数
                    example: 100
                  page:
                    type: number
                    description: 当前页
                    example: 1
                  limit:
                    type: number
                    description: 每页数量
                    example: 10
                  totalPages:
                    type: number
                    description: 总页数
                    example: 10
                required:
                  - data
                  - total
                  - page
                  - limit
                  - totalPages
        '401':
          description: 未授权
      tags:
        - 用户管理
      security:
        - bearer: []
...
  /users/{id}/reset-password:
    post:
      operationId: UsersController_resetPassword
      summary: 重置用户密码
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                password:
                  type: string
                  description: 新密码
                  example: newpassword
              required:
                - password
      responses:
        '200':
          description: 重置成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    description: 消息内容
                    example: 操作成功
                required:
                  - message
        '400':
          description: 参数错误
        '401':
          description: 未授权
        '403':
          description: 无权限
        '404':
          description: 用户不存在
      tags:
        - 用户管理
      security:
        - bearer: []
  /users/{id}/toggle-status:
    post:
      operationId: UsersController_toggleStatus
      summary: 切换用户状态
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: 切换成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    description: 消息内容
                    example: 操作成功
                required:
                  - message
        '401':
          description: 未授权
        '403':
          description: 无权限
        '404':
          description: 用户不存在
      tags:
        - 用户管理
      security:
        - bearer: []
...
info:
  title: OKR绩效考核系统API
  description: OKR绩效考核系统后端API文档
  version: '1.0'
  contact: {}
...
servers: []
components:
  securitySchemes:
    bearer:
      scheme: bearer
      bearerFormat: JWT
      type: http
...